EP-3399 #2 Split off ProjectedPolygons to own implementation file
