+++ openeo-geotrellis/src/main/scala/org/openeo/geotrellis/ComputeStatsGeotrellisAdapter.scala
@@ -1,7 +1,6 @@
 package org.openeo.geotrellis
 
 import java.io.File
-import java.net.{MalformedURLException, URL}
 import java.time.ZonedDateTime
 import java.time.format.DateTimeFormatter
 import java.util
@@ -10,126 +9,18 @@ import be.vito.eodata.extracttimeseries.geotrellis._
 import be.vito.eodata.geopysparkextensions.KerberizedAccumuloInstance
 import be.vito.eodata.processing.MaskedStatisticsProcessor.StatsMeanResult
 import geotrellis.layer._
-import geotrellis.proj4.{CRS, LatLng}
+import geotrellis.proj4.CRS
 import geotrellis.raster.histogram.Histogram
 import geotrellis.raster.summary.Statistics
 import geotrellis.raster.{FloatConstantNoDataCellType, UByteConstantNoDataCellType, UByteUserDefinedNoDataCellType}
 import geotrellis.spark._
 import geotrellis.store.accumulo.AccumuloInstance
-import geotrellis.vector.io.json.JsonFeatureCollection
-import geotrellis.vector.{Geometry, MultiPolygon, Polygon, _}
 import org.apache.spark.SparkContext
 import org.apache.spark.rdd.RDD
-import org.geotools.data.Query
-import org.geotools.data.shapefile.ShapefileDataStore
-import org.geotools.data.simple.SimpleFeatureIterator
 import org.slf4j.{Logger, LoggerFactory}
 
 import scala.collection.JavaConverters._
-import scala.io.Source
-import _root_.io.circe.DecodingFailure
 
-
-
-case class ProjectedPolygons(polygons: Array[MultiPolygon], crs: CRS)
-
-object ProjectedPolygons {
-  private type JList[T] = java.util.List[T]
-
-  def fromWkt(polygon_wkts: JList[String], polygons_srs: String): ProjectedPolygons = {
-    val polygons = polygon_wkts.asScala.map(parsePolygonWkt).toArray
-    val crs: CRS = CRS.fromName(polygons_srs)
-    ProjectedPolygons(polygons, crs)
-  }
-
-  def parsePolygonWkt(polygonWkt: String): MultiPolygon = {
-    val geometry: Geometry = polygonWkt.parseWKT()
-    geometry match {
-      case multiPolygon: MultiPolygon => multiPolygon
-      case _ => MultiPolygon(geometry.asInstanceOf[Polygon])
-    }
-  }
-
-  def fromVectorFile(vector_file: String): ProjectedPolygons = {
-    val vectorUrl = try {
-      new URL(vector_file)
-    } catch {
-      case _: MalformedURLException => new URL(s"file://$vector_file")
-    }
-
-    val filename = vectorUrl.getPath.split("/").last
-
-    if (filename.endsWith(".shp")) readSimpleFeatures(vectorUrl)
-    else readMultiPolygonsFromGeoJson(vectorUrl)
-  }
-
-  // adapted from Geotrellis' ShapeFileReader to avoid having too much in memory
-  private def readSimpleFeatures(shpUrl: URL): ProjectedPolygons = {
-    val ds = new ShapefileDataStore(shpUrl)
-    val ftItr: SimpleFeatureIterator = ds.getFeatureSource.getFeatures.features
-
-    try {
-      val featureCount = ds.getCount(Query.ALL)
-      require(featureCount < Int.MaxValue)
-
-      val simpleFeatures = new Array[MultiPolygon](featureCount.toInt)
-
-      for (i <- simpleFeatures.indices) {
-        val multiPolygon = ftItr.next().getAttribute(0) match {
-          case multiPolygon: MultiPolygon => multiPolygon
-          case polygon: Polygon => MultiPolygon(polygon)
-        }
-
-        simpleFeatures(i) = multiPolygon
-      }
-
-      // FIXME: read it from the shp and default to LatLng
-      ProjectedPolygons(simpleFeatures, LatLng)
-    } finally {
-      ftItr.close()
-      ds.dispose()
-    }
-  }
-
-  private def readMultiPolygonsFromGeoJson(geoJsonUrl: URL): ProjectedPolygons = {
-    // FIXME: stream it instead
-    val src = Source.fromURL(geoJsonUrl)
-
-    val multiPolygons = try {
-      val geoJson = src.mkString
-
-      def children(geometryCollection: GeometryCollection): Stream[Geometry] = {
-        def from(i: Int): Stream[Geometry] =
-          if (i >= geometryCollection.getNumGeometries) Stream.empty
-          else geometryCollection.getGeometryN(i) #:: from(i + 1)
-
-        from(0)
-      }
-
-      def asMultiPolygons(geometry: Geometry): Array[MultiPolygon] = geometry match {
-        case polygon: Polygon => Array(MultiPolygon(polygon))
-        case multiPolygon: MultiPolygon => Array(multiPolygon)
-        case geometryCollection: GeometryCollection => children(geometryCollection).map {
-          case polygon: Polygon => MultiPolygon(polygon)
-          case multiPolygon: MultiPolygon => multiPolygon
-        }.toArray
-      }
-
-      try {
-        asMultiPolygons(geoJson.parseGeoJson[Geometry]())
-      } catch {
-        case _: DecodingFailure =>
-          val featureCollection = geoJson.parseGeoJson[JsonFeatureCollection]()
-          featureCollection.getAllGeometries()
-            .flatMap(asMultiPolygons)
-            .toArray
-      }
-    } finally src.close()
-
-    ProjectedPolygons(multiPolygons, LatLng)
-  }
-
-}
 
 object ComputeStatsGeotrellisAdapter {
   private type JMap[K, V] = java.util.Map[K, V]
